Refactor : create의 반환타입 교체
